+++ CHANGES.md
@@ -3,6 +3,7 @@ Version 1.0.2 (pending)
 - Search staff list server-side for performance on slow clients
 - Added each user's total logged in time for the displayed day to the Login History report
 - Time period is displayed when you hover over a cell in the Login History Report
+- Added autocomplete dropdown to staff search
 
 Version 1.0.1 (11/14/2019)
 - Staff check in: don't search staff list until 2 characters have been entered in search (for performance)